lots of bug fixes. First simulation works

+++ ships/ship_basic.py
@@ -10,12 +10,27 @@ class Ship(Description,
         ListOfWares):
 
     def __init__(self):
+        # Description
         self.description_short = "Very basic ship class."
         self.description_full = self.description_short+" Not very useful to play with, but can be a basic for another classes."
+        # Coordinates
+        self.coordinates = [0,0]
+        # TimeDistanceStatistics
+        self.total_distance_traveling = 0
+        self.total_time_traveling = 0.0
+        # LastNextDestinations
+        self.last_destination = "" #port
+        self.next_destination = "" #port
+        self.time_reach = 0.0 #time when reaching the next destination
+        # ShipState
+        self.current_state = -1
+        # ListOfWares
+        self.wares = []
+        return None
 
     def get_next_destination_ship(self):
         # where the ship will go next
-        return get_next_destination_wares()
+        return self.get_next_destination_wares()
 
     def land_wares_to_port(self, wares_to_land, port, current_time):
         # method used to put correct wares to the port
@@ -24,6 +39,7 @@ class Ship(Description,
             ware.set_current_position(port)
             ware.update_ware_state(current_time)
             if not ware.read_destination_status():
+                # wares landed to port, but not in the correct place
                 wares_to_port.append(ware)
         return wares_to_port
 
@@ -31,6 +47,8 @@ class Ship(Description,
         # method used as docking the ship to the port
         land_port = self.read_next_destination()
         if land_port != port:
+            print "ERROR: ship: is on the wrong destination"
+            self.start_ship(current_time)
             return -1
         #add to all wares distance between last stop and current
         distance = self.distance_between_destinations() #find last trip length
@@ -39,7 +57,14 @@ class Ship(Description,
         self.set_coordinates(port.read_coordinates()) #update position of the ship
         #find wares connected to the port (destination = port)
         wares_to_land = self.find_wares_with_destination(port) #get wares
+        if len(wares_to_land)==0 and port.number_of_wares()==0:
+            print "ERROR: ship: nothing to land and empty port"
+            return self.set_current_state(0)
+        if self.number_of_wares()==0 and port.number_of_wares()==0:
+            print "ERROR: ship: empty ship is docking to empty port"
+            return self.set_current_state(0)
         wares_to_port = self.land_wares_to_port(wares_to_land, port, current_time) #put them to port
+        port.load_ware_from_ship(wares_to_port, current_time)
         self.delete_wares_from_list(wares_to_land) #delete from ship
         #stop the ship
         return self.set_current_state(0)
